// src/http/enums/http_method.enum.ts
export enum HttpMethod {
  Copy = 'COPY',
  Delete = 'DELETE',
  Get = 'GET',
  Head = 'HEAD',
  Lock = 'LOCK',
  Mkcol = 'MKCOL',
  Move = 'MOVE',
  Options = 'OPTIONS',
  Patch = 'PATCH',
  Post = 'POST',
  PropFind = 'PROPFIND',
  PropPatch = 'PROPPATCH',
  Put = 'PUT',
  Search = 'SEARCH',
  Trace = 'TRACE',
  Unlock = 'UNLOCK',
}